requirements.txt
django==2.2.14
requests==2.24.0
psutil==5.6.6
django-crontab==0.7.1
python-magic==0.4.18    
